+++ stroom-core-server/src/main/java/stroom/proxy/repo/FileWalker.java
@@ -1,4 +1,4 @@
-package stroom.streamtask.server;
+package stroom.proxy.repo;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;