+++ stroom-core-server/src/main/java/stroom/streamtask/server/ProxyAggregationExecutor.java
@@ -23,38 +23,19 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Component;
 import stroom.jobsystem.server.JobTrackedSchedule;
-import stroom.proxy.repo.StroomZipFile;
+import stroom.proxy.repo.FileSetProcessor;
+import stroom.proxy.repo.RepositoryProcessor;
 import stroom.proxy.repo.StroomZipRepository;
-import stroom.streamtask.server.FileWalker.FileFilter;
-import stroom.streamtask.server.FileWalker.FileProcessor;
 import stroom.task.server.ExecutorProvider;
 import stroom.task.server.TaskContext;
-import stroom.task.server.ThreadPoolImpl;
 import stroom.util.config.PropertyUtil;
-import stroom.util.date.DateUtil;
-import stroom.util.io.CloseableUtil;
-import stroom.util.io.FileUtil;
-import stroom.util.logging.LogExecutionTime;
 import stroom.util.shared.ModelStringUtil;
 import stroom.util.shared.Task;
-import stroom.util.shared.ThreadPool;
 import stroom.util.spring.StroomScope;
 import stroom.util.spring.StroomSimpleCronSchedule;
 
 import javax.inject.Inject;
 import javax.inject.Provider;
-import java.nio.file.Path;
-import java.nio.file.attribute.BasicFileAttributes;
-import java.util.Collections;
-import java.util.Comparator;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.Executor;
-import java.util.function.Consumer;
-import java.util.stream.Collectors;
 
 /**
  * <p>
@@ -69,28 +50,19 @@ public class ProxyAggregationExecutor {
     private static final int DEFAULT_MAX_AGGREGATION = 10000;
     private static final long DEFAULT_MAX_STREAM_SIZE = ModelStringUtil.parseIECByteSizeString("10G");
     private static final int DEFAULT_MAX_FILE_SCAN = 10000;
-    private static final FileFilter ZIP_FILTER = (file, attrs) -> file.toString().endsWith(StroomZipRepository.ZIP_EXTENSION);
 
     private final TaskContext taskContext;
-    private final ExecutorProvider executorProvider;
-    private final Provider<FileSetProcessor> fileSetProcessorProvider;
+    private final RepositoryProcessor repositoryProcessor;
 
-    private final int threadCount;
-    private final int maxFilesPerAggregate;
-    private final int maxConcurrentMappedFiles;
-    private final long maxUncompressedFileSize;
-
-    private final StroomZipRepository stroomZipRepository;
-    
     @Inject
-    public ProxyAggregationExecutor(final TaskContext taskContext,
-                                    final ExecutorProvider executorProvider,
-                                    final Provider<FileSetProcessor> fileSetProcessorProvider,
-                                    @Value("#{propertyConfigurer.getProperty('stroom.proxyDir')}") final String proxyDir,
-                                    @Value("#{propertyConfigurer.getProperty('stroom.proxyThreads')}") final String threadCount,
-                                    @Value("#{propertyConfigurer.getProperty('stroom.maxAggregation')}") final String maxFilesPerAggregate,
-                                    @Value("#{propertyConfigurer.getProperty('stroom.maxAggregationScan')}") final String maxConcurrentMappedFiles,
-                                    @Value("#{propertyConfigurer.getProperty('stroom.maxStreamSize')}") final String maxUncompressedFileSize) {
+    ProxyAggregationExecutor(final TaskContext taskContext,
+                             final ExecutorProvider executorProvider,
+                             final Provider<FileSetProcessor> fileSetProcessorProvider,
+                             @Value("#{propertyConfigurer.getProperty('stroom.proxyDir')}") final String proxyDir,
+                             @Value("#{propertyConfigurer.getProperty('stroom.proxyThreads')}") final String threadCount,
+                             @Value("#{propertyConfigurer.getProperty('stroom.maxAggregation')}") final String maxFilesPerAggregate,
+                             @Value("#{propertyConfigurer.getProperty('stroom.maxAggregationScan')}") final String maxConcurrentMappedFiles,
+                             @Value("#{propertyConfigurer.getProperty('stroom.maxStreamSize')}") final String maxUncompressedFileSize) {
         this(
                 taskContext,
                 executorProvider,
@@ -112,14 +84,14 @@ public class ProxyAggregationExecutor {
                              final int maxConcurrentMappedFiles,
                              final long maxUncompressedFileSize) {
         this.taskContext = taskContext;
-        this.executorProvider = executorProvider;
-        this.fileSetProcessorProvider = fileSetProcessorProvider;
-        this.threadCount = threadCount;
-        this.maxFilesPerAggregate = maxFilesPerAggregate;
-        this.maxConcurrentMappedFiles = maxConcurrentMappedFiles;
-        this.maxUncompressedFileSize = maxUncompressedFileSize;
-
-        stroomZipRepository = new StroomZipRepository(proxyDir);
+        this.repositoryProcessor = new RepositoryProcessor(taskContext,
+                executorProvider,
+                fileSetProcessorProvider,
+                proxyDir,
+                threadCount,
+                maxFilesPerAggregate,
+                maxConcurrentMappedFiles,
+                maxUncompressedFileSize);
     }
 
     @StroomSimpleCronSchedule(cron = "0,10,20,30,40,50 * *")
@@ -127,75 +99,10 @@ public class ProxyAggregationExecutor {
     public void exec(final Task<?> task) {
         if (!taskContext.isTerminated()) {
             try {
-                final LogExecutionTime logExecutionTime = new LogExecutionTime();
-                LOGGER.info("exec() - started");
-
-                taskContext.info("Aggregate started {}, maxFilesPerAggregate {}, maxConcurrentMappedFiles {}, maxUncompressedFileSize {}",
-                        DateUtil.createNormalDateTimeString(System.currentTimeMillis()),
-                        ModelStringUtil.formatCsv(maxFilesPerAggregate),
-                        ModelStringUtil.formatCsv(maxConcurrentMappedFiles),
-                        ModelStringUtil.formatIECByteSizeString(maxUncompressedFileSize));
-
-                process(stroomZipRepository);
-
-                LOGGER.info("exec() - completed in {}", logExecutionTime);
+                repositoryProcessor.process();
             } catch (final Exception e) {
                 LOGGER.error(e.getMessage(), e);
             }
-        }
-    }
-
-    /**
-     * Process a Stroom zip repository,
-     *
-     * @param stroomZipRepository The Stroom zip repository to process.
-     */
-    private void process(final StroomZipRepository stroomZipRepository) {
-        try {
-            if (LOGGER.isDebugEnabled()) {
-                LOGGER.debug("process() - Scanning " + stroomZipRepository.getRootDir());
-            }
-
-            // Scan all of the zip files in the repository so that we can map zip files to feeds.
-            final ErrorReceiver errorReceiver = (path, message) -> addErrorMessage(path, message, true);
-            final ZipInfoConsumer zipInfoConsumer = new ZipInfoConsumer(
-                    stroomZipRepository,
-                    maxFilesPerAggregate,
-                    maxConcurrentMappedFiles,
-                    maxUncompressedFileSize,
-                    errorReceiver,
-                    fileSetProcessorProvider,
-                    executorProvider,
-                    threadCount);
-            final ZipInfoExtractor zipInfoExtractor = new ZipInfoExtractor(errorReceiver);
-            final FileProcessorImpl fileProcessor = new FileProcessorImpl(
-                    zipInfoExtractor,
-                    zipInfoConsumer,
-                    taskContext,
-                    executorProvider,
-                    threadCount);
-            final FileWalker fileWalker = new FileWalker();
-            fileWalker.walk(stroomZipRepository.getRootDir(), ZIP_FILTER, fileProcessor, taskContext);
-
-            // Wait for the file processor to complete.
-            fileProcessor.await();
-
-            // Complete processing remaining file sets.
-            zipInfoConsumer.complete();
-
-            LOGGER.debug("Completed");
-        } catch (final Exception e) {
-            LOGGER.error(e.getMessage(), e);
-        }
-    }
-
-    private void addErrorMessage(final Path path, final String msg, final boolean bad) {
-        StroomZipFile stroomZipFile = null;
-        try {
-            stroomZipFile = new StroomZipFile(path);
-            stroomZipRepository.addErrorMessage(stroomZipFile, msg, bad);
-        } finally {
-            CloseableUtil.closeLogAndIgnoreException(stroomZipFile);
         }
     }
 
@@ -212,190 +119,5 @@ public class ProxyAggregationExecutor {
         }
 
         return value;
-    }
-
-    private static class ZipInfoConsumer implements Consumer<ZipInfo> {
-        private final StroomZipRepository stroomZipRepository;
-        private final int maxFilesPerAggregate;
-        private final int maxConcurrentMappedFiles;
-        private final long maxUncompressedFileSize;
-        private final ErrorReceiver errorReceiver;
-        private final Provider<FileSetProcessor> fileSetProcessorProvider;
-        private final Executor executor;
-
-        private final Map<String, FileSet> fileSetMap = new ConcurrentHashMap<>();
-        private final Set<CompletableFuture> futures = Collections.newSetFromMap(new ConcurrentHashMap<>());
-
-        private int totalMappedFiles;
-
-        ZipInfoConsumer(final StroomZipRepository stroomZipRepository,
-                        final int maxFilesPerAggregate,
-                        final int maxConcurrentMappedFiles,
-                        final long maxUncompressedFileSize,
-                        final ErrorReceiver errorReceiver,
-                        final Provider<FileSetProcessor> fileSetProcessorProvider,
-                        final ExecutorProvider executorProvider,
-                        final int threadCount) {
-            this.stroomZipRepository = stroomZipRepository;
-            this.maxFilesPerAggregate = maxFilesPerAggregate;
-            this.maxConcurrentMappedFiles = maxConcurrentMappedFiles;
-            this.maxUncompressedFileSize = maxUncompressedFileSize;
-            this.errorReceiver = errorReceiver;
-            this.fileSetProcessorProvider = fileSetProcessorProvider;
-
-            final ThreadPool threadPool = new ThreadPoolImpl("File Set Processor", 5, 0, threadCount, 2 * threadCount);
-            executor = executorProvider.getExecutor(threadPool);
-        }
-
-        @Override
-        public synchronized void accept(final ZipInfo zipInfo) {
-            final String feedName = zipInfo.getFeedName();
-            if (feedName == null || feedName.length() == 0) {
-                errorReceiver.onError(zipInfo.getPath(), "Unable to find feed in header??");
-
-            } else {
-                LOGGER.debug("{} belongs to feed {}", zipInfo.getPath(), feedName);
-
-                FileSet fileSet = fileSetMap.computeIfAbsent(feedName, k -> new FileSet(feedName));
-
-                // See if the file set will overflow if we add this file.
-                if (fileSet.getFiles().size() > 0 &&
-                        (fileSet.getTotalUncompressedFileSize() + zipInfo.getUncompressedSize() > maxUncompressedFileSize ||
-                                fileSet.getTotalZipEntryCount() + zipInfo.getZipEntryCount() > maxFilesPerAggregate)) {
-
-                    // Send the file set for processing.
-                    processFileSet(fileSet);
-
-                    // Create a new file set.
-                    fileSet = fileSetMap.computeIfAbsent(feedName, k -> new FileSet(feedName));
-                }
-
-                // The file set is not full so add the file.
-                fileSet.add(zipInfo);
-                totalMappedFiles++;
-
-                // If the file set is now full send it for processing.
-                if (fileSet.getTotalUncompressedFileSize() >= maxUncompressedFileSize ||
-                        fileSet.getTotalZipEntryCount() >= maxFilesPerAggregate) {
-
-                    // Send the file set for processing.
-                    processFileSet(fileSet);
-
-                } else {
-
-                    // If we have reached the maximum number of concurrent mapped files then we need to send the largest set for processing.
-                    if (totalMappedFiles >= maxConcurrentMappedFiles) {
-                        final List<FileSet> sortedList = fileSetMap
-                                .values()
-                                .stream()
-                                .sorted(Comparator.comparing(FileSet::getTotalUncompressedFileSize).reversed())
-                                .collect(Collectors.toList());
-
-                        // Remove the biggest file set from the map.
-                        final FileSet biggestFileSet = sortedList.get(0);
-
-                        // Send the file set for processing.
-                        processFileSet(biggestFileSet);
-                    }
-                }
-            }
-        }
-
-        private synchronized void processFileSet(final FileSet fileSet) {
-            // Remove the full file set.
-            fileSetMap.remove(fileSet.getFeed());
-
-            // Reduce the total number of files we have mapped.
-            totalMappedFiles -= fileSet.getFiles().size();
-
-            try {
-                final Runnable runnable = () -> {
-                    final FileSetProcessor fileSetProcessor = fileSetProcessorProvider.get();
-                    fileSetProcessor.process(stroomZipRepository, fileSet);
-                };
-                final CompletableFuture<Void> completableFuture = CompletableFuture.runAsync(runnable, executor);
-                completableFuture.thenAccept(r -> futures.remove(completableFuture));
-                futures.add(completableFuture);
-            } catch (final RuntimeException e) {
-                LOGGER.error(e.getMessage(), e);
-            }
-        }
-
-        public synchronized void complete() {
-            // Send all remaining file sets for processing.
-            fileSetMap.values().forEach(this::processFileSet);
-
-            try {
-                CompletableFuture.allOf(futures.toArray(new CompletableFuture[0])).join();
-            } catch (final RuntimeException e) {
-                LOGGER.error(e.getMessage(), e);
-                throw e;
-            }
-        }
-    }
-
-    private static class FileProcessorImpl implements FileProcessor {
-        private final ZipInfoExtractor zipInfoExtractor;
-        private final TaskContext taskContext;
-        private final Executor executor;
-        private final Consumer<ZipInfo> zipInfoConsumer;
-        private final Set<CompletableFuture> futures = Collections.newSetFromMap(new ConcurrentHashMap<>());
-
-        FileProcessorImpl(final ZipInfoExtractor zipInfoExtractor,
-                          final Consumer<ZipInfo> zipInfoConsumer,
-                          final TaskContext taskContext,
-                          final ExecutorProvider executorProvider,
-                          final int threadCount) {
-            this.zipInfoExtractor = zipInfoExtractor;
-            this.zipInfoConsumer = zipInfoConsumer;
-            this.taskContext = taskContext;
-
-            final ThreadPool fileInspectorThreadPool = new ThreadPoolImpl("Proxy File Inspection", 5, 0, threadCount, 2 * threadCount);
-            executor = executorProvider.getExecutor(fileInspectorThreadPool);
-        }
-
-        @Override
-        public void process(final Path file, final BasicFileAttributes attrs) {
-            try {
-                final Runnable runnable = () -> {
-                    // Process the file to extract ZipInfo
-                    taskContext.setName("Extract Zip Info");
-                    taskContext.info(FileUtil.getCanonicalPath(file));
-
-                    if (!taskContext.isTerminated()) {
-                        final ZipInfo zipInfo = zipInfoExtractor.extract(file, attrs);
-                        if (zipInfo != null) {
-                            if (LOGGER.isTraceEnabled()) {
-                                LOGGER.trace("Consume: " + zipInfo);
-                            }
-                            processZipInfo(zipInfo);
-                        }
-                    }
-                };
-                final CompletableFuture<Void> completableFuture = CompletableFuture.runAsync(runnable, executor);
-                completableFuture.thenAccept(r -> futures.remove(completableFuture));
-                futures.add(completableFuture);
-            } catch (final RuntimeException e) {
-                LOGGER.error(e.getMessage(), e);
-            }
-        }
-
-        /**
-         * This method is synchronised so that the zip info consumer can operate in a thread safe manner.
-         *
-         * @param zipInfo The zip info to process.
-         */
-        private void processZipInfo(final ZipInfo zipInfo) {
-            zipInfoConsumer.accept(zipInfo);
-        }
-
-        public void await() {
-            try {
-                CompletableFuture.allOf(futures.toArray(new CompletableFuture[0])).join();
-            } catch (final RuntimeException e) {
-                LOGGER.error(e.getMessage(), e);
-                throw e;
-            }
-        }
     }
 }