+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+* Issue **#382** : Large zip files are now broken apart prior to proxy aggregation.
+
 * Change start script to use absolute paths for jar, config and logs to distinguish stroom and proxy instances.
 
 
@@ -16,6 +18,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * Issue **#1116** : Changed the way tasks are executed to ensure thread pools expand to the maximum number of threads specified rather than just queueing all tasks and only providing core threads.
 
 * Remove full path from file in sha256 hash file release artifact.
+
 
 ## [v6.0-beta.29] - 2019-03-13